pyproject.toml
[project]
name = "mcp2cli"
version = "0.0.1"
description = "Turn any MCP server or OpenAPI spec into a CLI"
readme = "README.md"
license = "MIT"
authors = [
    { name = "Stephan Fitzpatrick", email = "stephan@knowsuchagency.com" }
]
requires-python = ">=3.10"
dependencies = [
    "httpx",
    "pyyaml",
]

[project.optional-dependencies]
mcp = ["mcp>=1.0"]
test = ["pytest", "pytest-asyncio", "mcp>=1.0"]

[project.scripts]
mcp2cli = "mcp2cli:main"

[build-system]
requires = ["uv_build>=0.9.5,<0.10.0"]
build-backend = "uv_build"
